Update agent_uninstall.ps1
Check if the service exists before trying to remove it.

+++ packaging/googet/agent_uninstall.ps1
@@ -12,8 +12,10 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-Stop-Service GCEAgent -Verbose
-& sc.exe delete GCEAgent
+if (Get-Service GCEAgent -ErrorAction SilentlyContinue) {
+    Stop-Service GCEAgent -Verbose
+    & sc.exe delete GCEAgent
+}
 
 $compat_manager = 'GCEWindowsCompatManager'
 $name = 'GCEAgentManager'
